[#183] typing - socket 모듈 타입 추가
