+++ backend/package.json
@@ -71,6 +71,7 @@
         "@types/mongoose-auto-increment": "^5.0.33",
         "@types/morgan": "^1.9.2",
         "@types/node": "^14.14.41",
+        "@types/socket.io": "^2.1.13",
         "@typescript-eslint/eslint-plugin": "^4.22.0",
         "@typescript-eslint/parser": "^4.22.0",
         "apidoc": "^0.20.0",
