+++ backend/src/init.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import * as database from './utils/database'
 import app from './app'
-import initialization from './utils/initialization'
+import { initialize } from './utils/initialization'
 import { initSocket } from './utils/socket'
 import elasticApmNode from 'elastic-apm-node'
 
@@ -57,7 +57,7 @@ export default async function initApp({
 
         initDatabase(DATABASE_URI)
 
-        await initialization.initialize()
+        await initialize()
         await initSocket(app, SOCKET_PORT)
 
         app.listen(PORT, () => console.log(`Listening on port ${PORT}`))
